[identity] Fix build warnings
